# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CRM module for Uralintern",
    "description": "API documentation for Uralintern CRM Module.\n\n**Контроллеры:**\n- **auth-controller**: Отвечает за регистрацию и авторизацию пользователей, включая регистрацию с реферальной ссылкой и обновление токенов.\n- **user-info-controller**: Управляет пользовательскими данными, включая обновление профиля, получение реферальных ссылок и список руководителей.\n- **event-controller**: Обеспечивает управление мероприятиями, включая их создание, обновление, скрытие, удаление, а также доступ к информации о мероприятиях.\n- **message-controller**: Управляет шаблонами сообщений для мероприятий, включая их получение и обновление.\n- **event-student-controller**: Обеспечивает управление заявками и участием студентов в мероприятиях, включая отправку заявок, их принятие/отклонение и управление студентами на мероприятии.\n- **event-curator-controller**: Управляет заявками и участием кураторов в мероприятиях, включая их принятие/отклонение и управление кураторами.\n- **notification-controller**: Управляет уведомлениями пользователей, связанными с их участием в мероприятиях.\n",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://localhost/api",
      "description": "kayakto server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "description": "Авторизация пользователя. Возвращает токены доступа и обновления.\nДоступен для всех пользователей.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход в систему.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Регистрация пока не работает",
        "tags": [
          "auth-controller"
        ],
        "operationId": "register",
        "description": "Регистрация нового пользователя. Доступно для всех.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно зарегистрирован.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/auth/register-with-token": {
      "post": {
        "summary": "Регистрация пока не работает",
        "tags": [
          "auth-controller"
        ],
        "operationId": "registerWithToken",
        "description": "Регистрация администратора по токену, полученному от администратора. ВНИМАНИЕ нельзя указывать компетенции и роль, в примере показано.",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequestManager"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "refresh",
        "description": "Обновление токена доступа по рефреш-токену.\nДоступен для всех пользователей с валидным рефреш-токеном.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Токены успешно обновлены.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDTO"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
...
    "/api/statuses/{id}": {
      "put": {
        "tags": [
          "status-controller"
        ],
        "summary": "Обновить статус",
        "description": "Обновляет название существующего статуса по ID.\nДоступен только для администраторов.\n",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID статуса для обновления"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Статус успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationStatus"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос"
          },
          "404": {
            "description": "Статус не найден"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      },
      "delete": {
        "tags": [
          "status-controller"
        ],
        "summary": "Удалить статус",
        "description": "Удаляет статус по ID.\nДоступен только для администраторов.\n",
        "operationId": "deleteStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID статуса для удаления"
          }
        ],
        "responses": {
          "200": {
            "description": "Статус успешно удален"
          },
          "404": {
            "description": "Статус не найден"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/api/statuses": {
      "post": {
        "tags": [
          "status-controller"
        ],
        "summary": "Создать новый статус",
        "description": "Создает новый статус в системе.\nДоступен только для администраторов.\n",
        "operationId": "createStatus",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Статус успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationStatus"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос (например, пустое название статуса)"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      },
      "get": {
        "tags": [
          "status-controller"
        ],
        "summary": "Получить все статусы",
        "description": "Возвращает список всех доступных статусов.\nДоступен для всех аутентифицированных пользователей.\n",
        "operationId": "getAllStatuses",
        "responses": {
          "200": {
            "description": "Успешный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationStatus"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
...
      "BadRequestResponse": {
        "description": "Bad Request, запрос неверный со стороны клиента",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid argument"
                }
              }
            }
          }
        }
      },
...
      "InternalServerErrorResponse": {
        "description": "Internal Server Error, запрос вызвал ошибку на сервере",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "ApplicationStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true,
            "description": "Уникальный идентификатор статуса"
          },
          "name": {
            "type": "string",
            "description": "Название статуса",
            "example": "В обработке"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 123
          },
          "firstName": {
            "type": "string",
            "example": "Николай"
          },
          "lastName": {
            "type": "string",
            "example": "Иванов"
          },
          "surname": {
            "type": "string",
            "example": "Евгеньевич"
          },
          "telegramUrl": {
            "type": "string",
            "example": "https://t.me/kayakto"
          },
          "vkUrl": {
            "type": "string",
            "example": "https://vk.com/kayakto"
          },
          "role_enum": {
            "type": "string",
            "enum": [
              "ADMIN",
              "CURATOR",
              "MANAGER",
              "STUDENT"
            ],
            "example": "STUDENT"
          },
          "competencies": {
            "type": "string",
            "example": "Spring Boot, Gradle, git, Java, Docker"
          }
        }
      },
...
      "UserRequest": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "sign",
          "telegramUrl",
          "vkUrl"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Николай"
          },
          "lastName": {
            "type": "string",
            "example": "Иванов"
          },
          "surname": {
            "type": "string",
            "example": "Евгеньевич"
          },
          "email": {
            "type": "string",
            "example": "mail@mail.ru"
          },
          "sign": {
            "type": "string",
            "example": "password"
          },
          "telegramUrl": {
            "type": "string",
            "example": "https://t.me/kayakto"
          },
          "vkUrl": {
            "type": "string",
            "example": "https://vk.com/kayakto"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "CURATOR",
              "MANAGER",
              "STUDENT"
            ],
            "example": "STUDENT"
          },
          "competencies": {
            "type": "string",
            "example": "Spring Boot, Gradle, git, Java, Docker"
          }
        }
      },
      "UserRequestManager": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "sign",
          "telegramUrl",
          "vkUrl"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Николай"
          },
          "lastName": {
            "type": "string",
            "example": "Иванов"
          },
          "surname": {
            "type": "string",
            "example": "Евгеньевич"
          },
          "email": {
            "type": "string",
            "example": "mail@mail.ru"
          },
          "sign": {
            "type": "string",
            "example": "password"
          },
          "telegramUrl": {
            "type": "string",
            "example": "https://t.me/kayakto"
          },
          "vkUrl": {
            "type": "string",
            "example": "https://vk.com/kayakto"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "mail123@mail.ru"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        }
      },
...
      "TokensDTO": {
        "type": "object",
        "properties": {
          "access": {
            "type": "string"
          },
          "refresh": {
            "type": "string"
          }
        }
      },
...
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refresh": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}